feat: break saveRecord up into legacy operations
