+++ packages/legacy-compat/src/legacy-network-handler/legacy-network-handler.ts
@@ -3,10 +3,13 @@ import { assert } from '@ember/debug';
 import { Promise, resolve } from 'rsvp';
 
 import { DEBUG } from '@ember-data/env';
+import { LOG_PAYLOADS } from '@ember-data/private-build-infra/debugging';
+import { DEPRECATE_V1_RECORD_DATA } from '@ember-data/private-build-infra/deprecations';
 import type { Handler, NextFn } from '@ember-data/request/-private/types';
 import type Store from '@ember-data/store';
 import type { Snapshot } from '@ember-data/store/-private';
 import type { StoreRequestContext } from '@ember-data/store/-private/cache-handler';
+import type ShimModelClass from '@ember-data/store/-private/legacy-model-support/shim-model-class';
 import type { Collection } from '@ember-data/store/-private/record-arrays/identifier-array';
 import type {
   CollectionResourceDocument,
@@ -15,12 +18,19 @@ import type {
 } from '@ember-data/types/q/ember-data-json-api';
 import type { StableExistingRecordIdentifier, StableRecordIdentifier } from '@ember-data/types/q/identifier';
 import type { AdapterPayload, MinimumAdapterInterface } from '@ember-data/types/q/minimum-adapter-interface';
+import type { MinimumSerializerInterface } from '@ember-data/types/q/minimum-serializer-interface';
+import type { JsonApiValidationError } from '@ember-data/types/q/record-data-json-api';
 
 import { guardDestroyedStore } from './common';
-import FetchManager from './fetch-manager';
+import FetchManager, { SaveOp } from './fetch-manager';
 import { assertIdentifierHasId } from './identifier-has-id';
 import { normalizeResponseHelper } from './serializer-response';
 import SnapshotRecordArray from './snapshot-record-array';
+
+type AdapterErrors = Error & { errors?: unknown[]; isAdapterError?: true; code?: string };
+type SerializerWithParseErrors = MinimumSerializerInterface & {
+  extractErrors?(store: Store, modelClass: ShimModelClass, error: AdapterErrors, recordId: string | null): unknown;
+};
 
 export const LegacyNetworkHandler: Handler = {
   request<T>(context: StoreRequestContext, next: NextFn<T>) {
@@ -35,6 +45,12 @@ export const LegacyNetworkHandler: Handler = {
     }
 
     switch (context.request.op) {
+      case 'updateRecord':
+        return saveRecord(context);
+      case 'deleteRecord':
+        return saveRecord(context);
+      case 'createRecord':
+        return saveRecord(context);
       case 'findRecord':
         return findRecord(context);
       case 'findAll':
@@ -48,6 +64,147 @@ export const LegacyNetworkHandler: Handler = {
     }
   },
 };
+
+function saveRecord<T>(context: StoreRequestContext): Promise<T> {
+  const { store, data, op: operation } = context.request;
+  const { options, record: identifier } = data as { record: StableRecordIdentifier; options: Record<string, unknown> };
+
+  const saveOptions = Object.assign(
+    { [SaveOp]: operation as 'updateRecord' | 'deleteRecord' | 'createRecord' },
+    options
+  );
+  const fetchManagerPromise = store._fetchManager.scheduleSave(identifier, saveOptions);
+
+  return fetchManagerPromise
+    .then((payload) => {
+      if (LOG_PAYLOADS) {
+        try {
+          let data: unknown = payload ? JSON.parse(JSON.stringify(payload)) : payload;
+          // eslint-disable-next-line no-console
+          console.log(`EmberData | Payload - ${operation!}`, data);
+        } catch (e) {
+          // eslint-disable-next-line no-console
+          console.log(`EmberData | Payload - ${operation!}`, payload);
+        }
+      }
+      /*
+      // TODO @runspired re-evaluate the below claim now that
+      // the save request pipeline is more streamlined.
+
+      Note to future spelunkers hoping to optimize.
+      We rely on this `run` to create a run loop if needed
+      that `store._push` and `store.saveRecord` will both share.
+
+      We use `join` because it is often the case that we
+      have an outer run loop available still from the first
+      call to `store._push`;
+     */
+      store._join(() => {
+        let data = payload && payload.data;
+        if (!data) {
+          assert(
+            `Your ${identifier.type} record was saved to the server, but the response does not have an id and no id has been set client side. Records must have ids. Please update the server response to provide an id in the response or generate the id on the client side either before saving the record or while normalizing the response.`,
+            identifier.id
+          );
+        }
+
+        const identifierCache = store.identifierCache;
+        let actualIdentifier = identifier;
+        if (operation !== 'deleteRecord' && data) {
+          actualIdentifier = identifierCache.updateRecordIdentifier(identifier, data);
+        }
+
+        //We first make sure the primary data has been updated
+        const cache = DEPRECATE_V1_RECORD_DATA ? store._instanceCache.getResourceCache(actualIdentifier) : store.cache;
+        cache.didCommit(identifier, data);
+
+        if (payload && payload.included) {
+          store._push({ data: null, included: payload.included });
+        }
+      });
+      return store.peekRecord(identifier);
+    })
+    .catch((e: unknown) => {
+      let err = e;
+      if (!e) {
+        err = new Error(`Unknown Error Occurred During Request`);
+      } else if (typeof e === 'string') {
+        err = new Error(e);
+      }
+      adapterDidInvalidate(store, identifier, err as Error);
+      throw err;
+    }) as Promise<T>;
+}
+
+function adapterDidInvalidate(
+  store: Store,
+  identifier: StableRecordIdentifier,
+  error: Error & { errors?: JsonApiValidationError[]; isAdapterError?: true; code?: string }
+) {
+  if (error && error.isAdapterError === true && error.code === 'InvalidError') {
+    let serializer = store.serializerFor(identifier.type) as SerializerWithParseErrors;
+
+    // TODO @deprecate extractErrors being called
+    // TODO remove extractErrors from the default serializers.
+    if (serializer && typeof serializer.extractErrors === 'function') {
+      let errorsHash = serializer.extractErrors(store, store.modelFor(identifier.type), error, identifier.id) as Record<
+        string,
+        string | string[]
+      >;
+      error.errors = errorsHashToArray(errorsHash);
+    }
+  }
+  const cache = DEPRECATE_V1_RECORD_DATA ? store._instanceCache.getResourceCache(identifier) : store.cache;
+
+  if (error.errors) {
+    assert(
+      `Expected the cache in use by resource ${String(
+        identifier
+      )} to have a getErrors(identifier) method for retreiving errors.`,
+      typeof cache.getErrors === 'function'
+    );
+
+    let jsonApiErrors: JsonApiValidationError[] = error.errors;
+    if (jsonApiErrors.length === 0) {
+      jsonApiErrors = [{ title: 'Invalid Error', detail: '', source: { pointer: '/data' } }];
+    }
+    cache.commitWasRejected(identifier, jsonApiErrors);
+  } else {
+    cache.commitWasRejected(identifier);
+  }
+}
+
+function makeArray<T>(value: T | T[]): T[] {
+  return Array.isArray(value) ? value : [value];
+}
+
+const PRIMARY_ATTRIBUTE_KEY = 'base';
+function errorsHashToArray(errors: Record<string, string | string[]>): JsonApiValidationError[] {
+  const out: JsonApiValidationError[] = [];
+
+  if (errors) {
+    Object.keys(errors).forEach((key) => {
+      let messages = makeArray(errors[key]);
+      for (let i = 0; i < messages.length; i++) {
+        let title = 'Invalid Attribute';
+        let pointer = `/data/attributes/${key}`;
+        if (key === PRIMARY_ATTRIBUTE_KEY) {
+          title = 'Invalid Document';
+          pointer = `/data`;
+        }
+        out.push({
+          title: title,
+          detail: messages[i],
+          source: {
+            pointer: pointer,
+          },
+        });
+      }
+    });
+  }
+
+  return out;
+}
 
 function findRecord<T>(context: StoreRequestContext): Promise<T> {
   const { store, data } = context.request;
