+++ packages/store/package.json
@@ -32,16 +32,12 @@
     "@ember-data/model": "workspace:4.12.0-alpha.9",
     "@ember-data/json-api": "workspace:4.12.0-alpha.9",
     "@ember-data/graph": "workspace:4.12.0-alpha.9",
-    "@ember-data/legacy-compat": "workspace:4.12.0-alpha.9",
     "@ember-data/tracking": "workspace:4.12.0-alpha.9",
     "@ember/string": "^3.0.1",
     "@glimmer/tracking": "^1.1.2"
   },
   "peerDependenciesMeta": {
     "@ember-data/json-api": {
-      "optional": true
-    },
-    "@ember-data/legacy-compat": {
       "optional": true
     },
     "@ember-data/graph": {
